requirement.txt
openai
python-dotenv
requests
edge-tts
matplotlib
